+++ biome.json
@@ -24,9 +24,6 @@
       "recommended": true,
       "suspicious": {
         "noUnknownAtRules": "off"
-      },
-      "style": {
-        "useConsistentCurlyBraces": "error"
       }
     },
     "domains": {
